Bone debug renderables (#73)
* Moving bone information to mesh

Moving the bone information to the mesh from the geometryprepass, this way we can allow the mesh to draw the debug primitives and pass the bone matrices to the shaders

* Fixing broken bone information

* Update mesh.cpp

Removing an import that doesn't exist yet

* Update uiRenderer.cpp

Renaming "No Animation" to "Rest Post" as it is more descriptive

* Adding instance rendering support to commandbuffer

Draw and DrawIndexed always rendered one instance, but adding an optional parameter for istanced rendering

* Bone axis renderable

Added a debug renderable for bone axis (currently always enabled until I hook up the ui)

Split up the actual rendering, debug rendering and added a new debug mode RenderNoDepthDebug, for debug renderings that should be rendered ontop of everything

* Update sceneRenderer.cpp

Fixing resize issue for depthless debug command buffer

* Update primitive.cpp

Fixing incorrect buffer being released when initializing

* Update boundingBox.cpp

slight improvements to boundingbox rendering

* bone joint rendering

More correct bone rendering.
Store the bonejoinrenderable as a unique pointer because not all models have animations.
Render one joint instance for each m_boneWorldTransform

* Adding bone axis rendering

Instanced rendering supported for orientationgizmo
Slight change of the orientationgizmo vertex ubo

* Reducing shader amount

Joining a lot of the skinning/instancing logic into a single shader, with some additions
This should hopefully simplify things a bit

* Update cmfcontent.cpp

Adding correct bounding box for meshless files.
We then use the skeletons, but if they are not available then we return a sphere with a 1 radius centered at 0,0,0

* Update primitive.h

Removing an unneeded primitive attribute

* BoneTransforms now occupy layoutbinding 1

Which means color information is stored in layoutbinding 2

* Model now owns skeletal information

moved skeletal and animations from mesh to model.
The model now handles updating skeletal animations and calculating world transforms. It also owns the debug renderables for bones.

Also fixed a visualization shader bug (where it isn't appIied)

* Refactoring the activeAnimationOwner to be a shared_ptr

So we don't need to look this up everytime we need it

* Add bone selection and highlighting to UI and renderer

- Added StateCollection for selected bones in AppState with callback support.
- Updated flatcolor shaders to use per-vertex color and selection state.
- Refactored bone/joint/bounding box rendering to use new PrimitiveEffects utilities.
- Added PrimitiveEffects module for common effect creation.
- Enhanced PrimitiveRenderable and GraphicsEffect for dynamic storage buffer management.
- Changed Buffer/BufferBuilder to use size_t for sizes and strides.
- Enabled bone selection in UI (multi-select with Ctrl) and propagate to renderer.
- Updated ModelRenderable and AnimationState to handle selection highlighting.
- Removed legacy code paths and improved code consistency.

* Add AnimationState class for skeleton animation control

Introduces the AnimationState class to manage animation state for a single skeleton, including pose updates, animation playback, and debug rendering of bones, joints, and axes. The class handles bone mapping, responds to AppState changes, and provides methods for updating and visualizing animation state.

* Fixing flatcolor shader

Removed the if statement as it will make the dFdy and dFdx behave funny

* Fixing linter errors

* fixing linter errors

* fixing a compilation error on macs

* fixing build errors on mac

* Fixing review comments

* Fixing a build error

not good to use an uninitiaIized accumulated bounding box to get Size and Center.
Also fixing another uninitiaIzed box in cmflcontent.cpp

* Update cmfcontent.cpp

formatting...

---------

+++ carbon-mesh-viewer/src/rendering/vulkan/buffer.h
@@ -21,7 +21,7 @@ public:
 	~Buffer();
 
 	void Release( const Renderer* renderer );
-	VkResult Initialize( const Renderer* renderer, BufferType type, const uint8_t* data, uint32_t size, uint32_t stride );
+	VkResult Initialize( const Renderer* renderer, BufferType type, const uint8_t* data, size_t size, size_t stride );
 
 	VkBuffer GetGpuBuffer() const;
 	VkDescriptorBufferInfo& GetDescriptorBufferInfo();
@@ -30,9 +30,9 @@ public:
 
 	void CopyFromStaging( VkCommandBuffer commandBuffer );
 	void ReleaseStaging( const Renderer* renderer );
-	VkResult SetData( const Renderer* renderer, const uint8_t* data, uint32_t size );
-	uint32_t size() const;
-	uint32_t stride() const;
+	VkResult SetData( const Renderer* renderer, const uint8_t* data, size_t size );
+	size_t size() const;
+	size_t stride() const;
 
 
 private:
@@ -40,8 +40,8 @@ private:
 	VkResult CreateStorageBuffer( const Renderer* renderer, const uint8_t* data );
 	VkDeviceMemory m_memory;
 	VkBuffer m_buffer;
-	uint32_t m_size;
-	uint32_t m_stride;
+	size_t m_size;
+	size_t m_stride;
 
 	VkDeviceMemory m_stagingMemory;
 	VkBuffer m_stagingBuffer;
@@ -54,5 +54,5 @@ private:
 
 namespace BufferBuilder
 {
-Buffer* Build( const Renderer* renderer, const uint8_t* data, uint32_t size, BufferType type, uint32_t stride );
+Buffer* Build( const Renderer* renderer, const uint8_t* data, size_t size, BufferType type, size_t stride );
 }
